pyproject.toml
[tool.poetry]
name = "spaceriders_apiv2"
version = "0.1.0"
description = ""
authors = ["Jordi <jordi.hoock@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
gunicorn = "^20.1.0"
python-decouple = "^3.6"
beanie = { git = "https://github.com/roman-right/beanie.git", tag = "1.11.1" }
web3 = "^5.29.1"
PyJWT = "^2.4.0"
slots = "^0.4.0"
experience = "^0.0.1"
emcache = "^0.6.1"
aioschedule = "^0.5.2"
python-json-logger = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
